backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
